refactor(calendar): modularize UI components and extract useCalendarEvents hook
